fix: change id type for serial and initialize with 0

+++ src/main/resources/db/migration/V1__planet.sql
@@ -1,5 +1,5 @@
 create table planet (
-    id BIGINT primary key,
+    id SERIAL primary key,
     max_x INTEGER not null,
     max_y INTEGER not null,
     name varchar(200) not null
